Using uv to install forks

+++ src/nrp_devtools/commands/ui/link_assets.py
@@ -62,12 +62,34 @@ def copy_assets_to_webpack_build_dir(config: OARepoConfig):
     for kind, source_file, target_file in tqdm(
         _list_copied_files(copied), desc="Linking assets and statics"
     ):
+        print(kind, source_file, target_file)
+
+        # sometimes, the "invenio assets create" creates symlinks. If this happens,
+        # remove the symlink as it will be replaced with a file copy to enable
+        # the watcher later on
+
+        if target_file.exists():
+            # find first symlink above
+            symlinked_file_or_dir = target_file
+
+            # assert that symlinked_file_or_dir is inside the invenio_instance_path
+            # this is to avoid removing files outside of the invenio_instance_path
+            # in case of bug or user tampering the watch.list.json
+            assert symlinked_file_or_dir.is_relative_to(config.invenio_instance_path)
+
+            while symlinked_file_or_dir != config.invenio_instance_path:
+                if symlinked_file_or_dir.is_symlink():
+                    # remove the symlinked file_or_dir
+                    symlinked_file_or_dir.unlink()
+                    break
+                symlinked_file_or_dir = symlinked_file_or_dir.parent
+            else:
+                target_file.unlink()
+
         target_file.parent.mkdir(parents=True, exist_ok=True)
-        try:
-            target_file.unlink()
-        except FileNotFoundError:
-            pass
+
         # copy source file to target file
+        # print(f"Copying {source_file} -> {target_file}")
         shutil.copy(source_file, target_file)
